% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{squares a single value or vector}
\usage{
square(x)
}
\arguments{
\item{x, }{numeric}
}
\description{
squares a single value or vector
}
\examples{
square(5)
square(c(1,2,3))

}
